数据编辑模块改为分离式 index为组装件
