+++ src/view/excel/components/stepC.vue
@@ -1,0 +1,9 @@
+<template>
+  <h1>步骤3</h1>
+</template>
+<script>
+export default {
+  name: 'stepThree',
+  cname: '导入广告'
+}
+</script>
